chore: depreciate components/search
